+++ petty-module/petty-basic-info-service/src/main/java/com/github/pettyfer/basic/basicinfoserver/entity/SystemMenu.java
@@ -1,10 +1,10 @@
 package com.github.pettyfer.basic.basicinfoserver.entity;
 
-import com.baomidou.mybatisplus.activerecord.Model;
-import com.baomidou.mybatisplus.annotations.TableField;
+import java.math.BigDecimal;
 import com.baomidou.mybatisplus.annotations.TableId;
+import com.baomidou.mybatisplus.annotations.TableField;
+import com.baomidou.mybatisplus.activerecord.Model;
 import com.baomidou.mybatisplus.annotations.TableName;
-
 import java.io.Serializable;
 
 /**
@@ -13,7 +13,7 @@ import java.io.Serializable;
  * </p>
  *
  * @author Pettyfer
- * @since 2018-02-27
+ * @since 2018-02-28
  */
 @TableName("system_menu")
 public class SystemMenu extends Model<SystemMenu> {
@@ -59,7 +59,7 @@ public class SystemMenu extends Model<SystemMenu> {
      * 是否启用 @枚举 0-否；1-是
      */
     @TableField("IS_SHOW")
-    private Integer isShow;
+    private BigDecimal isShow;
     /**
      * 按钮标识
      */
@@ -69,17 +69,17 @@ public class SystemMenu extends Model<SystemMenu> {
      * 类型 @枚举 1-菜单；2-按钮；3-数据权限
      */
     @TableField("MENU_TYPE")
-    private Integer menuType;
+    private BigDecimal menuType;
     /**
      * 是否隐藏@枚举 0-否；1-是
      */
     @TableField("IS_HIDDEN")
-    private Integer isHidden;
+    private BigDecimal isHidden;
     /**
      * 是否第三方@枚举 0-否；1-是
      */
     @TableField("IS_EXTERNAL")
-    private Integer isExternal;
+    private BigDecimal isExternal;
     /**
      * 备注
      */
@@ -89,7 +89,7 @@ public class SystemMenu extends Model<SystemMenu> {
      * 菜单/按钮排序号
      */
     @TableField("SORT_NO")
-    private Integer sortNo;
+    private BigDecimal sortNo;
     /**
      * 创建人
      */
@@ -114,7 +114,7 @@ public class SystemMenu extends Model<SystemMenu> {
      * 删除标记  @枚举 0-否；1-是
      */
     @TableField("DEL_FLAG")
-    private Integer delFlag;
+    private BigDecimal delFlag;
 
 
     public String getMenuId() {
@@ -173,11 +173,11 @@ public class SystemMenu extends Model<SystemMenu> {
         this.icon = icon;
     }
 
-    public Integer getIsShow() {
+    public BigDecimal getIsShow() {
         return isShow;
     }
 
-    public void setIsShow(Integer isShow) {
+    public void setIsShow(BigDecimal isShow) {
         this.isShow = isShow;
     }
 
@@ -189,27 +189,27 @@ public class SystemMenu extends Model<SystemMenu> {
         this.permission = permission;
     }
 
-    public Integer getMenuType() {
+    public BigDecimal getMenuType() {
         return menuType;
     }
 
-    public void setMenuType(Integer menuType) {
+    public void setMenuType(BigDecimal menuType) {
         this.menuType = menuType;
     }
 
-    public Integer getIsHidden() {
+    public BigDecimal getIsHidden() {
         return isHidden;
     }
 
-    public void setIsHidden(Integer isHidden) {
+    public void setIsHidden(BigDecimal isHidden) {
         this.isHidden = isHidden;
     }
 
-    public Integer getIsExternal() {
+    public BigDecimal getIsExternal() {
         return isExternal;
     }
 
-    public void setIsExternal(Integer isExternal) {
+    public void setIsExternal(BigDecimal isExternal) {
         this.isExternal = isExternal;
     }
 
@@ -221,11 +221,11 @@ public class SystemMenu extends Model<SystemMenu> {
         this.remark = remark;
     }
 
-    public Integer getSortNo() {
+    public BigDecimal getSortNo() {
         return sortNo;
     }
 
-    public void setSortNo(Integer sortNo) {
+    public void setSortNo(BigDecimal sortNo) {
         this.sortNo = sortNo;
     }
 
@@ -261,11 +261,11 @@ public class SystemMenu extends Model<SystemMenu> {
         this.modifyDate = modifyDate;
     }
 
-    public Integer getDelFlag() {
+    public BigDecimal getDelFlag() {
         return delFlag;
     }
 
-    public void setDelFlag(Integer delFlag) {
+    public void setDelFlag(BigDecimal delFlag) {
         this.delFlag = delFlag;
     }
 
@@ -277,25 +277,25 @@ public class SystemMenu extends Model<SystemMenu> {
     @Override
     public String toString() {
         return "SystemMenu{" +
-                ", menuId=" + menuId +
-                ", code=" + code +
-                ", parentCode=" + parentCode +
-                ", menuName=" + menuName +
-                ", href=" + href +
-                ", target=" + target +
-                ", icon=" + icon +
-                ", isShow=" + isShow +
-                ", permission=" + permission +
-                ", menuType=" + menuType +
-                ", isHidden=" + isHidden +
-                ", isExternal=" + isExternal +
-                ", remark=" + remark +
-                ", sortNo=" + sortNo +
-                ", creator=" + creator +
-                ", createDate=" + createDate +
-                ", modifier=" + modifier +
-                ", modifyDate=" + modifyDate +
-                ", delFlag=" + delFlag +
-                "}";
+        ", menuId=" + menuId +
+        ", code=" + code +
+        ", parentCode=" + parentCode +
+        ", menuName=" + menuName +
+        ", href=" + href +
+        ", target=" + target +
+        ", icon=" + icon +
+        ", isShow=" + isShow +
+        ", permission=" + permission +
+        ", menuType=" + menuType +
+        ", isHidden=" + isHidden +
+        ", isExternal=" + isExternal +
+        ", remark=" + remark +
+        ", sortNo=" + sortNo +
+        ", creator=" + creator +
+        ", createDate=" + createDate +
+        ", modifier=" + modifier +
+        ", modifyDate=" + modifyDate +
+        ", delFlag=" + delFlag +
+        "}";
     }
 }
